Move AMO whitelist functions to taar_utils.py
