+++ mozetl/taar/taar_utils.py
@@ -1,4 +1,14 @@
 import boto3
+import json
+import logging
+
+from botocore.exceptions import ClientError
+
+logging.basicConfig(level=logging.INFO)
+logger = logging.getLogger(__name__)
+
+AMO_DUMP_BUCKET = 'telemetry-parquet'
+AMO_DUMP_KEY = 'telemetry-ml/addon_recommender/addons_database.json'
 
 
 def write_to_s3(source_file_name, s3_dest_file_name, s3_prefix, bucket):
@@ -41,3 +51,34 @@ def store_json_to_s3(json_data, base_filename, date, prefix, bucket):
     # Store a copy of the current JSON with datestamp.
     write_to_s3(FULL_FILENAME, archived_file_copy, prefix, bucket)
     write_to_s3(FULL_FILENAME, FULL_FILENAME, prefix, bucket)
+
+
+def load_amo_external_whitelist():
+    """ Download and parse the AMO add-on whitelist.
+
+    :raises RuntimeError: the AMO whitelist file cannot be downloaded or contains
+                          no valid add-ons.
+    """
+    final_whitelist = []
+    amo_dump = {}
+    try:
+        # Load the most current AMO dump JSON resource.
+        s3 = boto3.client('s3')
+        s3_contents = s3.get_object(Bucket=AMO_DUMP_BUCKET, Key=AMO_DUMP_KEY)
+        amo_dump = json.loads(s3_contents['Body'].read())
+    except ClientError:
+        logger.exception("Failed to download from S3", extra={
+            "bucket": AMO_DUMP_BUCKET,
+            "key": AMO_DUMP_KEY})
+
+    # If the load fails, we will have an empty whitelist, this may be problematic.
+    for key, value in amo_dump.items():
+        addon_files = value.get('current_version', {}).get('files', {})
+        # If any of the addon files are web_extensions compatible, it can be recommended.
+        if any([f.get("is_webextension", False) for f in addon_files]):
+            final_whitelist.append(value['guid'])
+
+    if len(final_whitelist) == 0:
+        raise RuntimeError("Empty AMO whitelist detected")
+
+    return final_whitelist
